login is working without juser

+++ src/libraries/default/sessions/storage/database.php
@@ -2,6 +2,9 @@
 
 class LibSessionsStorageDatabase extends LibSessionsStorageAbstract
 {
+    /**
+    *   @param hold session data
+    */
     private $_data = null;
 
 	/**
@@ -14,7 +17,7 @@ class LibSessionsStorageDatabase extends LibSessionsStorageAbstract
 	 */
 	function open($save_path, $session_name)
 	{
-		return true;
+        return true;
 	}
 
 	/**
@@ -40,9 +43,11 @@ class LibSessionsStorageDatabase extends LibSessionsStorageAbstract
 	{
         $this->_data = '';
 
-        $session = KService::get('repos:sessions.session')
-        ->getQuery()
-        ->fetch(array('id' => $id));
+        if ($id == '') {
+            return $this->_data;
+        }
+
+        $session = KService::get('repos:sessions.session')->find(array('id' => $id));
 
         if (isset($session)) {
             $this->_data = (string) $session->meta;
@@ -59,40 +64,41 @@ class LibSessionsStorageDatabase extends LibSessionsStorageAbstract
 	 * @param string $session_data  The session data.
 	 * @return boolean  True on success, false otherwise.
 	 */
-	function write($id = '', $meta = '')
+	function write($id, $session_data)
 	{
-        $session = KService::get('repos:sessions.session')
-        ->getQuery()
-        ->fetch(array('id' => $id));
+        if ($id == '' && $session_data == '') {
+            return false;
+        }
 
-        if(isset($session)) {
+        $session = KService::get('repos:sessions.session')->find(array('id' => $id));
+
+        if (isset($session)) {
 
             $session
-            ->set('id', $id)
-            ->set('meta', $meta)
+            ->set('meta', $session_data)
             ->save();
-
-            $this->_data = $meta;
 
         } else {
 
-            $session = KService::get('repos:sessions.session')->getEntity();
-            $session
+            KService::get('repos:sessions.session')
+            ->getEntity()
             ->set('id', $id)
-            ->set('meta', '')
+            ->set('meta', $session_data)
             ->save();
-
-            $this->_data = '';
         }
+
+        $this->_data = $session_data;
 
 		return true;
 	}
 
     function update($id) {
 
-        $session = KService::get('repos:sessions.session')
-        ->getQuery()
-        ->fetch(array('id' => $id));
+        if ($id == '') {
+            return false;
+        }
+
+        $session = KService::get('repos:sessions.session')->find(array('id' => $id));
 
         if(isset($session)) {
             return $session->set('time', time())->save();
@@ -111,9 +117,11 @@ class LibSessionsStorageDatabase extends LibSessionsStorageAbstract
 	  */
 	function destroy($id)
 	{
-        $session = KService::get('repos:sessions.session')
-        ->getQuery()
-        ->fetch(array('id' => $id));
+        if ($id == '') {
+            return false;
+        }
+
+        $session = KService::get('repos:sessions.session')->find(array('id' => $id));
 
         if (isset($session)) {
 		    $session->delete();
